requirements.txt
chronos-forecasting
pandas
matplotlib
numpy
torch
plotly
colorama
flask
